+++ src/main/java/CSE4186/interview/controller/LoginController.java
@@ -3,8 +3,10 @@ package CSE4186.interview.controller;
 
 import CSE4186.interview.controller.dto.BaseResponseDto;
 import CSE4186.interview.controller.dto.UserDTO;
+import CSE4186.interview.login.Oauth2UserService;
 import CSE4186.interview.service.UserService;
 import CSE4186.interview.utils.ApiUtil;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletRequest;
@@ -28,6 +30,7 @@ import java.util.stream.Collectors;
 public class LoginController {
 
     private final UserService userService;
+    private final Oauth2UserService OAuth2UserService;
 
     @PostMapping("/join")
     @Operation(summary = "Join User", description = "회원가입")
@@ -74,5 +77,20 @@ public class LoginController {
                         exceptionCode.equals(null)?"":exceptionCode,
                         ""
                 ));
+    }
+
+    @PostMapping("oauth2/google")
+    BaseResponseDto<Map<String,String>> getOauth2Token(@RequestBody UserDTO.oauth2LoginRequest oauth2LoginRequest, HttpServletRequest httpServletRequest) throws JsonProcessingException {
+
+        //1. 코드를 받는다.
+        String code= oauth2LoginRequest.getCode();
+
+        //2. 코드를 사용해 토큰을 받아온다.
+        String token= OAuth2UserService.requestGoogleToken(code);
+
+        //2. 토큰을 통해 구글 서버에서 계정 정보를 가져온다.
+        BaseResponseDto<Map<String,String>> response= OAuth2UserService.requestGoogleAccountAndLogin(token);
+
+        return response;
     }
 }
